refactor: move PackageVersion to common package
