+++ packages/common/src/debmagic/common/package_version.py
